+ (client) terminal UI + (server) timestamp convert to UTC
