+++ internal/client/run.go
@@ -5,6 +5,8 @@ import (
 
 	"github.com/MikeRez0/gophkeeper/internal/adapter/config"
 	"github.com/MikeRez0/gophkeeper/internal/adapter/logger"
+	"github.com/MikeRez0/gophkeeper/internal/client/app"
+	"github.com/MikeRez0/gophkeeper/internal/client/tui"
 	"go.uber.org/zap"
 )
 
@@ -16,7 +18,7 @@ func Run() error {
 
 	log := logger.NewLogger(conf.App)
 
-	app, err := NewApp(conf, log)
+	app, err := app.NewApp(conf, log)
 	if err != nil {
 		log.Error("error creating client app", zap.Error(err))
 		return err
@@ -32,38 +34,54 @@ func Run() error {
 		return err
 	}
 
-	kc := app.Keychains[0]
-	// item := kc.NewItem(domain.KCItemTypePassword)
-	// item.SetLabel("my site pass")
-	// item.MetaDataSetItem(domain.KCMetaKeyComment, "my test comment")
-	// item.MetaDataSetItem(domain.KCMetaKeyLogin, "admin")
-	// item.MetaDataSetItem(domain.KCMetaKeySite, "google.com")
+	for _, k := range app.Keychains {
+		if err = app.SyncKeychain(k); err != nil {
+			log.Error("error syncing keychain", zap.Error(err))
+			return err
+		}
+	}
 
-	kc.Pass = "test1"
-	// p := "mysuper-puper-password"
+	c, err := tui.NewUIController(app, log)
+	if err != nil {
+		log.Error("error creating TUI", zap.Error(err))
+		return err
+	}
+
+	err = c.Run()
+	if err != nil {
+		log.Error("error running app", zap.Error(err))
+		return err
+	}
+	// kc := app.Keychains[0]
+	// item := kc.NewItem(domain.KCItemTypeString)
+	// item.SetLabel("my string")
+	// item.MetaDataSetItem(domain.KCMetaKeyComment, "my comment to string")
+
+	// kc.Pass = "test1"
+	// p := "don't tell what i did last summer"
 	// err = kc.StoreSecret(item, []byte(p))
 	// if err != nil {
 	// 	log.Error("error storing secret", zap.Error(err))
 	// 	return err
 	// }
 
-	for i, k := range app.Keychains {
-		if err = app.SyncKeychain(k); err != nil {
-			log.Error("error syncing keychain", zap.Error(err))
-			return err
-		}
+	// for i, k := range app.Keychains {
+	// 	if err = app.SyncKeychain(k); err != nil {
+	// 		log.Error("error syncing keychain", zap.Error(err))
+	// 		return err
+	// 	}
 
-		fmt.Printf("Keychain %v: %v \n", i, k)
+	// 	fmt.Printf("Keychain %v: %v \n", i, k)
 
-		for j, item := range k.Items {
-			s, err := k.GetSecret(item)
-			if err != nil {
-				log.Error("error reading secret", zap.Error(err))
-			}
+	// 	for j, item := range k.Items {
+	// 		s, err := k.GetSecret(item)
+	// 		if err != nil {
+	// 			log.Error("error reading secret", zap.Error(err))
+	// 		}
 
-			fmt.Printf("%d: %v\n", j, string(s))
-		}
-	}
+	// 		fmt.Printf("%d: %v\n", j, string(s))
+	// 	}
+	// }
 
 	return nil
 }
